src/main/requirements.txt
annoy==1.17.0
cycler==0.10.0
decorator==4.4.2
easydict==1.9
faiss-gpu==1.7.0
imageio==2.9.0
install==1.3.4
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.1.2
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
tqdm==4.59.0
umap-learn==0.5.1
